fixed spleef
