+++ src/main/java/org/braekpo1nt/mctmanager/games/game/spleef/SpleefGame.java
@@ -11,7 +11,6 @@ import org.braekpo1nt.mctmanager.games.game.interfaces.MCTGame;
 import org.braekpo1nt.mctmanager.games.game.spleef.config.SpleefConfig;
 import org.braekpo1nt.mctmanager.games.game.spleef.config.SpleefConfigController;
 import org.braekpo1nt.mctmanager.games.utils.ParticipantInitializer;
-import org.braekpo1nt.mctmanager.ui.TimeStringUtils;
 import org.braekpo1nt.mctmanager.ui.sidebar.Headerable;
 import org.braekpo1nt.mctmanager.ui.sidebar.KeyLine;
 import org.braekpo1nt.mctmanager.ui.sidebar.Sidebar;
@@ -20,11 +19,9 @@ import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.HandlerList;
 import org.bukkit.event.Listener;
-import org.bukkit.event.entity.EntityDamageEvent;
 import org.bukkit.event.entity.FoodLevelChangeEvent;
 import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.event.player.PlayerDropItemEvent;
-import org.bukkit.scheduler.BukkitRunnable;
 import org.bukkit.scoreboard.Scoreboard;
 import org.bukkit.scoreboard.Team;
 
@@ -43,8 +40,6 @@ public class SpleefGame implements Listener, MCTGame, Configurable, Headerable {
     private List<Player> admins = new ArrayList<>();
     private List<SpleefRound> rounds = new ArrayList<>();
     private int currentRoundIndex = 0;
-    private int descriptionPeriodTaskId;
-    private boolean descriptionShowing = false;
     private boolean gameActive = false;
     
     public SpleefGame(Main plugin, GameManager gameManager) {
@@ -82,13 +77,14 @@ public class SpleefGame implements Listener, MCTGame, Configurable, Headerable {
         for (int i = 0; i < config.getRounds(); i++) {
             rounds.add(new SpleefRound(plugin, gameManager, this, config, sidebar, adminSidebar));
         }
+        rounds.get(0).setFirstRound(true);
         plugin.getLogger().info(String.format("rounds.size() = %d", rounds.size()));
         currentRoundIndex = 0;
         setupTeamOptions();
         startAdmins(newAdmins);
         displayDescription();
         gameActive = true;
-        startDescriptionPeriod();
+        startNextRound();
         Bukkit.getLogger().info("Started Spleef");
     }
     
@@ -150,7 +146,6 @@ public class SpleefGame implements Listener, MCTGame, Configurable, Headerable {
             }
         }
         rounds.clear();
-        descriptionShowing = false;
         gameActive = false;
         for (Player participant : participants) {
             resetParticipant(participant);
@@ -177,31 +172,6 @@ public class SpleefGame implements Listener, MCTGame, Configurable, Headerable {
     
     private void resetAdmin(Player admin) {
         adminSidebar.removePlayer(admin);
-    }
-    
-    @EventHandler
-    public void onPlayerDamage(EntityDamageEvent event) {
-        if (!gameActive) {
-            return;
-        }
-        if (event.getCause().equals(EntityDamageEvent.DamageCause.VOID)) {
-            return;
-        }
-        if (!(event.getEntity() instanceof Player participant)) {
-            return;
-        }
-        if (!participants.contains(participant)) {
-            return;
-        }
-        if (descriptionShowing) {
-            event.setCancelled(true);
-            return;
-        }
-        if (currentRoundIndex < 0 || rounds.size() <= currentRoundIndex) {
-            return;
-        }
-        SpleefRound round = rounds.get(currentRoundIndex);
-        round.onPlayerDamage(participant, event);
     }
     
     @EventHandler
@@ -262,9 +232,6 @@ public class SpleefGame implements Listener, MCTGame, Configurable, Headerable {
                 new KeyLine("title", title),
                 new KeyLine("round", String.format("Round %d/%d", currentRoundIndex+1, config.getRounds()))
         );
-        if (descriptionShowing) {
-            return;
-        }
         if (currentRoundIndex < rounds.size()) {
             SpleefRound currentRound = rounds.get(currentRoundIndex);
             if (currentRound.isActive()) {
@@ -280,9 +247,6 @@ public class SpleefGame implements Listener, MCTGame, Configurable, Headerable {
         }
         resetParticipant(participant);
         participants.remove(participant);
-        if (descriptionShowing) {
-            return;
-        }
         if (currentRoundIndex < rounds.size()) {
             SpleefRound currentRound = rounds.get(currentRoundIndex);
             if (currentRound.isActive()) {
@@ -291,28 +255,6 @@ public class SpleefGame implements Listener, MCTGame, Configurable, Headerable {
         }
     }
     
-    private void startDescriptionPeriod() {
-        descriptionShowing = true;
-        this.descriptionPeriodTaskId = new BukkitRunnable() {
-            private int count = config.getDescriptionDuration();
-            @Override
-            public void run() {
-                if (count <= 0) {
-                    sidebar.updateLine("timer", "");
-                    adminSidebar.updateLine("timer", "");
-                    descriptionShowing = false;
-                    startNextRound();
-                    this.cancel();
-                    return;
-                }
-                String timeLeft = TimeStringUtils.getTimeString(count);
-                String timerString = String.format("Starting soon: %s", timeLeft);
-                sidebar.updateLine("timer", timerString);
-                adminSidebar.updateLine("timer", timerString);
-                count--;
-            }
-        }.runTaskTimer(plugin, 0L, 20L).getTaskId();
-    }
     
     public void roundIsOver() {
         if (currentRoundIndex+1 >= rounds.size()) {
@@ -332,7 +274,7 @@ public class SpleefGame implements Listener, MCTGame, Configurable, Headerable {
     }
     
     private void cancelAllTasks() {
-        Bukkit.getScheduler().cancelTask(descriptionPeriodTaskId);
+        
     }
     
     private void initializeAdminSidebar() {
